### 2023-12-21 #6165 @editorialbot
[PRE REVIEW]: SelfEEG: A Python library for Self-Supervised Learning in Electroencephalography
**Submitting author:** @fedepup (<a href="http://orcid.org/0009-0004-0698-962X">Federico Del Pup</a>)
**Repository:** https://github.com/MedMaxLab/selfEEG
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
