### 2023-12-24 129fc80
updated notebooks and augmentation benchmark
