+++ .github/workflows/python-app.yml
@@ -1,4 +1,4 @@
-name: Python package
+name: Unittest
 
 on:
   push:
