### 2023-12-31 e717f91
updated SSL survey bib entry
