+++ Paper/bibliography.bib
@@ -33,13 +33,15 @@
 }
 
 
-@article{DelPup2023,
-author = {Del Pup, Federico and Atzori, Manfredo},
-title = {Applications of Self-Supervised Learning to Biomedical Signals: where are we now},
-year = {2023},
-month = {4},
-doi = {10.36227/techrxiv.22567021.v2}
-}
+@article{delpup2023,
+  author={Del Pup, Federico and Atzori, Manfredo},
+  journal={IEEE Access}, 
+  title={Applications of Self-Supervised Learning to Biomedical Signals: a Survey}, 
+  year={2023},
+  volume={11},
+  number={},
+  pages={144180-144203},
+  doi={10.1109/ACCESS.2023.3344531}}
 
 
 @article{pytorch,
