+++ Paper/bibliography.bib
@@ -33,7 +33,7 @@
 }
 
 
-@article{delpup2023,
+@article{DelPup2023,
   author={Del Pup, Federico and Atzori, Manfredo},
   journal={IEEE Access}, 
   title={Applications of Self-Supervised Learning to Biomedical Signals: a Survey}, 
