### 2024-01-07 811356f
solved problems with non integer window length
