### 2024-01-08 #6165 @emdupre
👋 Hi @fedepup, and thanks for your submission to JOSS !

If you have suggestions for potential reviewers, please let me know by listing their names or GitHub handles (without the @, so they don't receive a notification) here.

You can suggest reviewers from any relevant project, though we often recommend starting with [this database](https://reviewers.joss.theoj.org/) of people who have already agreed to review for JOSS.

I'll put together a list of folks to reach out to as reviewers based on these suggestions and my own recommendations. I'll update this thread with that outreach.
