### 2024-01-08 #6165 @fedepup
Hi @emdupre, thanks a lot for having decided to take care of our submission. Due to the waitlist, I was expecting the review to be on pause for a month or two, so I'm really happy you speeded up the process so much.

Regarding potential reviewers, one of my colleagues (and co-author) posted about our submission on Mastodon ([link](https://neuromatch.social/@fabrice13/111619278051636658)) and it seems that the following researchers showed interest in trying the library (although I don't know about reviewing it):
- behinger (listed in the database)
- jkanev (not listed in the database but seems really interested)

In addition, I tried to search for potential reviewers in the JOSS database, using as filters Python as the programming language and Artificial Intelligence (especially deep learning) and neuroscience (especially time series like EEG) as topic areas. I came up with the following names:
- wmvanvliet
- ntolley
- Bsingstad

Hope this will help you. If you have any other suggestions based on your expertise, or you don't feel my list matches the topics of selfEEG (deep learning + EEG), please feel free to give priority to your names.

I'm new to this, but I'm really excited about this type of review, which I hope will go as smoothly as possible. From my side, rest assured that my colleagues and I will give priority to any raised issues and try to resolve them as fast as possible.
