### 2024-01-08 #6165 @emdupre
Hi @fedepup, and thanks for these review suggestions ! I'll consider them alongside my own, and I'll aim to send out invitations within the next day or two.

Thank you, too, for your quick response and engagement ! I know that JOSS's unique review process will be unfamiliar on first submission, but I hope I can help it unfold as smoothly and transparently as possible. If you have any questions during the review process, of course, please don't hesitate to ask !
