### 2024-01-11 #6165 @emdupre
👋 Hi @Bsingstad,

Would you be willing to review _SelfEEG: A Python library for Self-Supervised Learning in Electroencephalography_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/MedMaxLab/selfEEG) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.06165/joss.06165/10.21105.joss.06165.pdf

If you're available to review this work within the next month—or six weeks, at most—please let me know and I'll add you as a reviewer.

Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
