### 2024-01-12 #6165 @fedepup
@wmvanvliet, @vferat - Thank you very much for accepting to review our work! Hope @Bsingstad will be able to join as well. As already said, I will give priority to any of your concerns in order to speed up the review process.

@behinger - Thank you, too, for the interest shown both here and on Mastodon. Hope you will get better soon.

@emdupre - Just to be transparent, I submitted a slightly modified version of the JOSS preprint to arXiv few weeks ago (I added a note linking to the JOSS preprint policy and clearly stating that the article was submitted to JOSS for possible publication). Since JOSS articles are short by nature, arXiv put the preprint "on hold", but in the end it was finally accepted yesterday ([preprint link](https://arxiv.org/abs/2401.05405)). Hope this will help to promote the library as well as this wonderful journal. I will add a link to the preprint in the official repo, which I will of course remove if the article will be considered worthy of being published by JOSS.
