### 2024-01-12 #6165 @emdupre
> @emdupre - Just to be transparent, I submitted a slightly modified version of the JOSS preprint to arXiv few weeks ago (I added a note linking to the JOSS preprint policy and clearly stating that the article was submitted to JOSS for possible publication). Since JOSS articles are short by nature, arXiv put the preprint "on hold", but in the end it was finally accepted yesterday ([preprint link](https://arxiv.org/abs/2401.05405)). Hope this will help to promote the library as well as this wonderful journal. I will add a link to the preprint in the official repo, which I will of course remove if the article will be considered worthy of being published by JOSS.

Thank you for letting me know re : the arXiv link, @fedepup ! Just confirming that it's fine to have a preprint posted of this submission. If _SelfEEG_ is accepted for publication in JOSS, you won't need to remove this preprint ; please just make sure to update the appropriate metadata on arXiv !
