+++ .readthedocs.yaml
@@ -10,6 +10,18 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.11"
+  jobs:
+		post_checkout:
+			# Cancel building pull requests when there aren't pushed changes in:
+			# - the docs directory
+			# - the selfeeg directory
+			# - the .readthedocs.yaml fle
+			# see (https://docs.readthedocs.io/en/stable/build-customization.html#cancel-build-based-on-a-condition)
+			- |
+				if [ "$READTHEDOCS_VERSION_TYPE" = "external" ] && git diff --quiet origin/main -- docs/ selfeeg/ .readthedocs.yaml;
+				then
+					exit 183;
+				fi
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
