### 2024-01-13 4154e64
added name and surname field in JOSS paper
