### 2024-01-13 6465279
Update .readthedocs.yaml
Tried to cancel build when no changes in the docs or selfeeg folder are pushed internally. If this pass I will add readthedocs.yaml in the git diff list
