+++ .readthedocs.yaml
@@ -22,6 +22,11 @@ build:
         then
           exit 183;
         fi
+      - |
+        if git diff --quiet origin/main -- docs/ selfeeg/ ;
+        then
+          exit 183;
+        fi
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
