### 2024-01-13 423f534
Update .readthedocs.yaml
Confirmed the decision to leave exit 183 condition and remove docs badge on readme
