### 2024-01-19 MedMaxLab/selfEEG#1 @wmvanvliet
Copyright holder in LICENSE.md
In `LICENSE.md`, the copyright holder's name is "fedepup", a Github username. Perhaps that could be changed to a real person's or team's name.
