### 2024-01-19 MedMaxLab/selfEEG#2 @wmvanvliet
pup_fede_cnd conda channel not found?
When attempting to install the package through conda, I get an error that the following URL cannot be resolved:
`https://conda.anaconda.org/pup_fede_cnd/win-64/current_repodata.json`

Does that mean that the `pup_fede_cnd` channel does not exist?
