### 2024-01-19 MedMaxLab/selfEEG#2 @fedepup
Sorry, there was a little typo in the command. The channel is "pup_fede_Cnd" . This is the reference link to the [anaconda](https://anaconda.org/Pup_Fede_Cnd/selfeeg) page. I will also change the description there. Thank you for having noticed the typo. Let me know if this solve the issue.
