### 2024-01-19 MedMaxLab/selfEEG#1 @fedepup
I've changed the copyright holder's name from "fedepup" to "MedMax Team".  Hope this will solve the issue, otherwise I will change it again with my full name (other team members are ok with that).
