### 2024-01-19 e3351b4
Corrected wrong conda installation command
