+++ README.md
@@ -81,7 +81,7 @@ pip install selfeeg[interactive]
 
 SelfEEG can be also installed via conda by running the following command:   
 ```
-conda install -c pup_fede_cnd -c pytorch selfeeg
+conda install -c Pup_Fede_Cnd -c pytorch selfeeg
 ```
 
 **Good practice**
