### 2024-01-19 1ba8cd0
Update meta.yaml
The description reported a wrong conda install command.
