+++ conda_recipe/meta.yaml
@@ -42,13 +42,13 @@ about:
   description: |
     To install selfEEG in a fresh environment, run the following command:
     
-    ``conda install -c pup_fede_cnd -c pytorch selfeeg``
+    ``conda install -c Pup_Fede_Cnd -c pytorch selfeeg``
     
     Alternatively, we suggest to first install PyTorch, Torchvision and Torchaudio by choosing the 
     right configuration depending on your preferences (see the official PyTorch documentation); 
     then, run the following command:
     
-    ``conda install -c pup_fede_cnd selfeeg``
+    ``conda install -c Pup_Fede_Cnd selfeeg``
     
     .
   doc_url: https://selfeeg.readthedocs.io/en/latest/index.html
