### 2024-01-22 8289f92
Make selfeeg compatible with python<3.11
