### 2024-01-22 e8de590
Some spelling/grammar tweaks to the paper
