### 2024-01-22 #6224 @wmvanvliet
Esxcellent work, @fedepup and team! I think this package will definitely help with the development of self-supervised learning techniques on EEG data.
