### 2024-01-22 MedMaxLab/selfEEG#2 @wmvanvliet
Ah, it's some weird issue with mamba. When I use plain conda, it indeed works.
