### 2024-01-22 MedMaxLab/selfEEG#6 @vferat
[JOSS] Review 
Hey all,

I'm opening this issue to allow me to make comments for the JOSS review (openjournals/joss-reviews#6224).

While testing the library:

- [x] https://github.com/MedMaxLab/selfEEG/blob/947b08d79e7597bd11b5907497aadd2fe85333fb/selfeeg/dataloading/load.py#L306
Got and error while working with `freq `of type `float`: `Unknown format code 'd' for object of type 'float'`
I can also recommend using [logging](https://docs.python.org/3/howto/logging.html) instead of `print `for verbose.

## Documentation

- [x]  I recommend providing a "real world" examplein the tutorial section. I think most people who will used the library will be familiar with [MNE-python](https://github.com/mne-tools/mne-python), therefore It would be interested to provide an notebook using a MNE dataset ( such as the [eegbci dataset](https://mne.tools/stable/generated/mne.datasets.eegbci.load_data.html#mne.datasets.eegbci.load_data)) to demonstrate use case of the library.

## Errors

I have tried the library on 3 different computers. For two of them, I encounter the same error while running the `SSL_guide `notebook
```
Empty                                     Traceback (most recent call last)
File [d:\Python\envs\selfEEG\Lib\site-packages\torch\utils\data\dataloader.py:1133](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1133), in _MultiProcessingDataLoaderIter._try_get_data(self, timeout)
   [1132](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1132) try:
-> [1133](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1133)     data = self._data_queue.get(timeout=timeout)
   [1134](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1134)     return (True, data)

File [d:\Python\envs\selfEEG\Lib\multiprocessing\queues.py:114](file:///D:/Python/envs/selfEEG/Lib/multiprocessing/queues.py:114), in Queue.get(self, block, timeout)
    [113](file:///D:/Python/envs/selfEEG/Lib/multiprocessing/queues.py:113)     if not self._poll(timeout):
--> [114](file:///D:/Python/envs/selfEEG/Lib/multiprocessing/queues.py:114)         raise Empty
    [115](file:///D:/Python/envs/selfEEG/Lib/multiprocessing/queues.py:115) elif not self._poll():

Empty:

RuntimeError                              Traceback (most recent call last)
Cell In[8], [line 1](vscode-notebook-cell:?execution_count=8&line=1)
----> [1](vscode-notebook-cell:?execution_count=8&line=1) loss_info = SelfMdl.fit(train_dataloader = trainloader, augmenter=Augmenter, epochs=5,
      [2](vscode-notebook-cell:?execution_count=8&line=2)                         optimizer=optimizer, loss_func= loss, loss_args= loss_arg,
      [3](vscode-notebook-cell:?execution_count=8&line=3)                         lr_scheduler= scheduler, EarlyStopper=earlystop,
      [4](vscode-notebook-cell:?execution_count=8&line=4)                         validation_dataloader=valloader,
      [5](vscode-notebook-cell:?execution_count=8&line=5)                         verbose=True, device= device, return_loss_info=True
      [6](vscode-notebook-cell:?execution_count=8&line=6)                        )

File [~\Documents\GitHub\selfEEG\selfeeg\ssl\ssl.py:1042](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1042), in SimCLR.fit(self, train_dataloader, epochs, optimizer, augmenter, loss_func, loss_args, lr_scheduler, EarlyStopper, validation_dataloader, verbose, device, cat_augmentations, return_loss_info)
   [1036](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1036)     self.train()
   [1038](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1038) with tqdm.tqdm(total=N_train+N_val, ncols=100,
   [1039](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1039)            bar_format='{desc}{percentage:3.0f}%|{bar:15}| {n_fmt}/{total_fmt}'
   [1040](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1040)            ' [{rate_fmt}{postfix}]',
   [1041](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1041)            disable=not(verbose), unit=' Batch',file=sys.stdout) as pbar:
-> [1042](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1042)     for batch_idx, X in enumerate(train_dataloader):
   [1044](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1044)         optimizer.zero_grad()
   [1046](https://file+.vscode-resource.vscode-cdn.net/c%3A/Users/Victor/Documents/GitHub/selfEEG/Notebooks/~/Documents/GitHub/selfEEG/selfeeg/ssl/ssl.py:1046)         if X.device.type!=device.type:

File [d:\Python\envs\selfEEG\Lib\site-packages\torch\utils\data\dataloader.py:631](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:631), in _BaseDataLoaderIter.__next__(self)
    [628](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:628) if self._sampler_iter is None:
    [629](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:629)     # TODO(https://github.com/pytorch/pytorch/issues/76750)
    [630](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:630)     self._reset()  # type: ignore[call-arg]
--> [631](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:631) data = self._next_data()
    [632](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:632) self._num_yielded += 1
    [633](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:633) if self._dataset_kind == _DatasetKind.Iterable and \
    [634](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:634)         self._IterableDataset_len_called is not None and \
    [635](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:635)         self._num_yielded > self._IterableDataset_len_called:

File [d:\Python\envs\selfEEG\Lib\site-packages\torch\utils\data\dataloader.py:1329](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1329), in _MultiProcessingDataLoaderIter._next_data(self)
   [1326](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1326)     return self._process_data(data)
   [1328](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1328) assert not self._shutdown and self._tasks_outstanding > 0
-> [1329](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1329) idx, data = self._get_data()
   [1330](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1330) self._tasks_outstanding -= 1
   [1331](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1331) if self._dataset_kind == _DatasetKind.Iterable:
   [1332](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1332)     # Check for _IterableDatasetStopIteration

File [d:\Python\envs\selfEEG\Lib\site-packages\torch\utils\data\dataloader.py:1295](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1295), in _MultiProcessingDataLoaderIter._get_data(self)
   [1291](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1291)     # In this case, `self._data_queue` is a `queue.Queue`,. But we don't
   [1292](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1292)     # need to call `.task_done()` because we don't use `.join()`.
   [1293](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1293) else:
   [1294](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1294)     while True:
-> [1295](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1295)         success, data = self._try_get_data()
   [1296](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1296)         if success:
   [1297](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1297)             return data

File [d:\Python\envs\selfEEG\Lib\site-packages\torch\utils\data\dataloader.py:1146](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1146), in _MultiProcessingDataLoaderIter._try_get_data(self, timeout)
   [1144](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1144) if len(failed_workers) > 0:
   [1145](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1145)     pids_str = ', '.join(str(w.pid) for w in failed_workers)
-> [1146](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1146)     raise RuntimeError(f'DataLoader worker (pid(s) {pids_str}) exited unexpectedly') from e
   [1147](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1147) if isinstance(e, queue.Empty):
   [1148](file:///D:/Python/envs/selfEEG/Lib/site-packages/torch/utils/data/dataloader.py:1148)     return (False, None)

RuntimeError: DataLoader worker (pid(s) 2916, 11640, 12340, 7664) exited unexpectedly
```
Everything was working perfectly on the 3rd one.

## Code base

This section is not required per say by the JOSS review guidelines, but seems to me to be an important point. I therefore leave you free to address these points or not, and will not delay the review for that reason.

If you want the library to evolve over time, I suggest:
- [x]  to add additional tests by installing the "main" version of https://github.com/pytorch/pytorch so as to be able to anticipate the changes needed to remain compatible.
- [x]  use a test matrix to check compatibility with several OS/python versions
- [x]  to maintain a simple & clean code base, add codestyle tests or a precommit (example: https://github.com/vferat/pycrostates/blob/main/.pre-commit-config.yaml)

## Conclusion
`selfEEG`  is a functional python library that and should make it easy to apply SSL models on EEG data. The code base is solid. Thanks for all your hard work
