+++ selfeeg/dataloading/load.py
@@ -1559,7 +1559,7 @@ class EEGDataset(Dataset):
             start=(self.Nsample-round(self.Nsample*self.overlap))*(partition - cumidx)
             end=start+self.Nsample
             if end>self.currEEG.shape[-1]: # in case of partial ending samples
-                sample=self.currEEG[(*first_dims_idx,slice(None),slice(-self.Nsample))]
+                sample= self.currEEG[(*first_dims_idx,slice(None),slice(self.currEEG.shape[-1]-Nsample, self.currEEG.shape[-1]))]
             else:
                 sample=self.currEEG[(*first_dims_idx,slice(None),slice(start, end))]
         else:
