### 2024-01-22 MedMaxLab/selfEEG#6 @vferat
Hey @fedepup,

No problem, I was indeed using the PIP version, I'll switch to `main`  to avoid outdated comments. Thanks for the quick reply !
