### 2024-01-22 MedMaxLab/selfEEG#6 @fedepup
> No problem, I was indeed using the PIP version, I'll switch to main to avoid outdated comments.

Thank you and sorry again for the inconvenience.

> BTW, if you are interested, you can find a GitHub workflow to automatically push a Github release to PYPI [here](https://github.com/vferat/pycrostates/blob/main/.github/workflows/publish.yaml) (just make sure to change package the version)

I will certainly add it in my to-do list!
