### 2024-02-03 1548188
conf.py will retrieve version from txt
