+++ docs/conf.py
@@ -17,10 +17,16 @@ copyright = '2023, MedMax Team'
 author = 'MedMax Team'
 
 
+version_path = os.path.abspath(os.path.join('..', 'selfeeg',
+                             'VERSION.txt'))
+
+with open(version_path, 'r') as fd:
+    version = fd.read().rstrip()
+
 # The short X.Y version
-version = '0.1.0'
+version = version
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = version
 
 # -- General configuration ---------------------------------------------------
 
