### 2024-02-03 MedMaxLab/selfEEG#6 @vferat
Hey @fedepup,
I've updated my review (first message). Only minor comments, mainly adding  a more concrete example in the tutorial section.
Otherwise, everything looks good to me !
Nice work !
