### 2024-02-03 MedMaxLab/selfEEG#6 @fedepup
Thanks for the update. I will add another notebook with a more concrete example as soon as possible.

For the error in the `SSL_guide` notebook, I think you have run the one in the `Notebooks` folder, right? The one in the `docs` folder (which I considered the "official" example) has some minor updates based on wmvanvliet revision. In short, I think the error is related to the `num_worker` parameter of the pytorch Dataloaders. May I ask you if you can try to run the one in the `docs` folder or set the parameter num_worker=0 in the one you have encountered the error? From my side, if this solves the issue, I will copy the notebooks of the `docs` folder inside the `notebooks` one so to keep everything updated. (I've run the `SSL_guide` in the docs folder notebook with a mac, linux and windows devices and everything worked fine.)

For the last three points, I thinks it's not a problem to work on that, but I will do it after the notebook concerns.

Thank you again!
