### 2024-02-05 e573df1
updated Notebooks and added new example
