### 2024-02-06 94ef392
additional improvements on docs plus expanded gitignore
