+++ .github/workflows/python-app.yml
@@ -3,7 +3,7 @@ name: Unittest
 on:
   push:
     paths:
-    - 'selfeeg/**'
+    - 'selfeeg/**' 
     - 'test/**'
     - '.github/workflows/**'
 
