### 2024-02-06 MedMaxLab/selfEEG#6 @vferat
Hey @fedepup !

I really like the new tutorial and thing I will rely help people not familiar with SSL to have a good start !

I keep this issue open for the maintenance part, but everything LGTM ! Thanks for all the hard work and for  providing this software to our community !
