### 2024-02-06 #6224 @vferat
Review completed !

Thanks to the authors for their responsiveness! The library delivers on its promise to simplify access to EEG data
All my comments have been addressed and I can only recommend to accept this publication.

Thank to all authors for making this tool available to the community !
