+++ .github/workflows/python-app.yml
@@ -10,11 +10,11 @@ on:
 jobs:
   selfeegtest:
 
-    runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.11"]
-    
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ["3.10", "3.11"]
+    runs-on: ${{ matrix.os }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
