+++ selfeeg/augmentation/compose.py
@@ -9,12 +9,12 @@ __all__ = ['DynamicSingleAug','RandomAug','SequentialAug','StaticSingleAug']
 
 class StaticSingleAug():
     """
-    ``StaticSingleAug`` performs a single data augmentation 
-    where the optional arguments are previously set and given during initialization. 
-    No random choice of the arguments is performed. The class accepts multiple set of 
+    ``StaticSingleAug`` performs a single data augmentation
+    where the optional arguments are previously set and given during initialization.
+    No random choice of the arguments is performed. The class accepts multiple set of
     optional arguments. In this case they are called individually at each class
-    call, in a sequential and cyclic manner. This means that the first call uses the first set of 
-    arguments, the second will use the second set of arguments, and so on. 
+    call, in a sequential and cyclic manner. This means that the first call uses the first set of
+    arguments, the second will use the second set of arguments, and so on.
     When the last set of arguments is used, the class will restart from the
     first set of arguments.
 
@@ -24,29 +24,29 @@ class StaticSingleAug():
     Parameters
     ----------
     augmentation: function
-        The augmentation function to apply. It can be a custom function, 
+        The augmentation function to apply. It can be a custom function,
         but the first argument must be the element to augment.
     arguments: list, dict, list[list or dict], optional
         The set of arguments to pass to the augmentation function. It can be:
-        
-            1. None. In this case the default parameters of the function are used. 
-            
-            2. list. In this case the function is called with the sintax 
+
+            1. None. In this case the default parameters of the function are used.
+
+            2. list. In this case the function is called with the sintax
             ``augmentation(x, *arguments)``
-            
-            3. dict. In this case the function is called with the sintax 
+
+            3. dict. In this case the function is called with the sintax
             ``augmentation(x, **arguments)``
-            
-            4. a list of dicts or lists. This is a particular case where multiple combinations 
-            of arguments are given. Each element of the list must be a list or a dict 
-            with the specific argument combination. Every time ``PerformAugmentation`` is called, 
+
+            4. a list of dicts or lists. This is a particular case where multiple combinations
+            of arguments are given. Each element of the list must be a list or a dict
+            with the specific argument combination. Every time ``PerformAugmentation`` is called,
             one of the given combinations is used to perform the data augmentation. The list
-            is followed sequentially with repetition, meaning that the first call uses the first 
-            set of arguments of the list, the second call uses the second set of arguments, and so on. 
+            is followed sequentially with repetition, meaning that the first call uses the first
+            set of arguments of the list, the second call uses the second set of arguments, and so on.
             When the last element of the list is used, the function will restart scrolling
-            the given list. 
-        
-        Default = None 
+            the given list.
+
+        Default = None
 
     Methods
     -------
@@ -59,7 +59,7 @@ class StaticSingleAug():
     >>> import torch
     >>> BatchEEG = torch.zeros(16,32,1024) + torch.sin(torch.linspace(0, 8*np.pi,1024))
     >>> Fs = 64
-    >>> Aug_eye = aug.StaticSingleAug(aug.add_eeg_artifact, 
+    >>> Aug_eye = aug.StaticSingleAug(aug.add_eeg_artifact,
     ...                               [{'Fs': Fs, 'artifact': 'eye', 'amplitude': 0.5, 'batch_equal': False},
     ...                                {'Fs': Fs, 'artifact': 'eye', 'amplitude': 1.0, 'batch_equal': False}
     ...                               ]
@@ -68,7 +68,7 @@ class StaticSingleAug():
[…]
