+++ selfeeg/models/zoo.py
@@ -34,21 +34,21 @@ class DepthwiseConv2d(nn.Conv2d):
         The depth multiplier. Output channels will be depth_multiplier*in_channels.
     kernel_size: int or tuple
         Size of the convolving kernel.
-    stride: int or tuple, optional 
-        Stride of the convolution. 
-        
+    stride: int or tuple, optional
+        Stride of the convolution.
+
         Default = 1
-    padding: int, tuple or str, optional 
+    padding: int, tuple or str, optional
         Padding added to all four sides of the input.
-        
+
         Default = 0
-    dilation: int or tuple, optional 
-        Spacing between kernel elements. 
-        
+    dilation: int or tuple, optional
+        Spacing between kernel elements.
+
         Default = 1
-    bias: bool, optional 
-        If True, adds a learnable bias to the output. 
-        
+    bias: bool, optional
+        If True, adds a learnable bias to the output.
+
         Default = True
     max_norm: float, optional
         The maximum norm each filter can have. If None no constraint will be included.
@@ -64,14 +64,14 @@ class DepthwiseConv2d(nn.Conv2d):
     >>> out = mdl(x)
     >>> print(out.shape) # shoud return torch.Size([4, 2, 8, 64])
     >>> print(torch.isnan(out).sum()) # shoud return 0
-    
+
     """
-    def __init__(self, in_channels, depth_multiplier, kernel_size, 
+    def __init__(self, in_channels, depth_multiplier, kernel_size,
                  stride=1, padding='same', dilation=1, bias=False, max_norm=None
                 ):
-        super(DepthwiseConv2d, self).__init__(in_channels, depth_multiplier*in_channels, 
+        super(DepthwiseConv2d, self).__init__(in_channels, depth_multiplier*in_channels,
                                               kernel_size, groups=in_channels,
-                                              stride=stride, padding=padding, 
+                                              stride=stride, padding=padding,
                                               dilation=dilation, bias=bias)
         if max_norm is not None:
             if max_norm <=0:
@@ -85,14 +85,14 @@ class DepthwiseConv2d(nn.Conv2d):
     def scale_norm(self, eps=1e-9):
         """
         Citing the Tensorflow documentation, the implementation tries to replicate this
-        
-        integer, axis along which to calculate weight norms. 
-        For instance, in a Dense layer the weight 
-        matrix has shape (input_dim, output_dim), set axis to 0 to constrain each 
+
+        integer, axis along which to calculate weight norms.
+        For instance, in a Dense layer the weight
+        matrix has shape (input_dim, output_dim), set axis to 0 to constrain each
         weight vector of length (input_dim,). In a Conv2D layer with
-        data_format="channels_last", the weight tensor has 
-        shape (rows, cols, input_depth, output_depth), set axis to [0, 1, 2] 
-        to constrain the weights 
+        data_format="channels_last", the weight tensor has
+        shape (rows, cols, input_depth, output_depth), set axis to [0, 1, 2]
+        to constrain the weights
         of each filter tensor of size (rows, cols, input_depth).
 
         :meta private:
@@ -107,7 +107,7 @@ class DepthwiseConv2d(nn.Conv2d):
         if (norm>self.max_norm).sum()>0:
             desired = torch.clamp(norm, 0, self.max_norm)
             self.weight = torch.nn.Parameter(self.weight*desired/ (eps + norm))
-    
+
     def forward(self, input):
         """
         :meta private:
@@ -115,11 +115,11 @@ class DepthwiseConv2d(nn.Conv2d):
         if self.max_norm is not None:
             self.scale_norm(self.max_norm)
[…]
