+++ test/EEGself/dataloading/load_test.py
@@ -10,317 +10,528 @@ import torch
 from torch.utils.data import DataLoader
 from selfeeg import dataloading as dl
 
+
 class TestDataloading(unittest.TestCase):
 
-    def create_dataset(self, folder_name='Simulated_EEG', data_len=[1024,4097], seed=1234):
-        N=1000
-        if not(os.path.isdir(folder_name)):
+    def create_dataset(self, folder_name="Simulated_EEG", data_len=[1024, 4097], seed=1234):
+        N = 1000
+        if not (os.path.isdir(folder_name)):
             os.mkdir(folder_name)
         np.random.seed(seed=seed)
         for i in range(N):
-            x = np.random.randn(2,np.random.randint(data_len[0],data_len[1]))
-            y = np.random.randint(1,5)
-            sample = {'data': x, 'label': y}
-            A, B, C = (int(i//200)+1), (int( (i - 200*int(i//200)))//5+1), (i%5+1)
-            file_name = 'Simulated_EEG/' + str(A) + '_' + str(B) + '_' + str(C) + '_1.pickle'
-            with open(file_name, 'wb') as f:
+            x = np.random.randn(2, np.random.randint(data_len[0], data_len[1]))
+            y = np.random.randint(1, 5)
+            sample = {"data": x, "label": y}
+            A, B, C = (int(i // 200) + 1), (int((i - 200 * int(i // 200))) // 5 + 1), (i % 5 + 1)
+            file_name = "Simulated_EEG/" + str(A) + "_" + str(B) + "_" + str(C) + "_1.pickle"
+            with open(file_name, "wb") as f:
                 pickle.dump(sample, f)
 
     def loadEEG(self, path, return_label=False):
-        with open(path, 'rb') as handle:
+        with open(path, "rb") as handle:
             EEG = pickle.load(handle)
-        x = EEG['data']
-        y = EEG['label']
+        x = EEG["data"]
+        y = EEG["label"]
         if return_label:
             return x, y
         else:
             return x
 
     def transformEEG(self, EEG, value=64):
-        EEG = EEG[:,:-value]
+        EEG = EEG[:, :-value]
         return EEG
 
     def makeGrid(self, pars_dict):
-        keys=pars_dict.keys()
-        combinations=itertools.product(*pars_dict.values())
-        ds=[dict(zip(keys,cc)) for cc in combinations]
+        keys = pars_dict.keys()
+        combinations = itertools.product(*pars_dict.values())
+        ds = [dict(zip(keys, cc)) for cc in combinations]
         return ds
 
     @classmethod
     def setUpClass(cls):
-        print('\n---------------------------')
-        print('TESTING DATALOADING MODULE')
-        print('the process will generetate two additional folders')
-        print('---------------------------')
-        if not(os.path.isdir('tmpsave')):
-                os.mkdir('tmpsave')
-        cls.eegpath  = 'Simulated_EEG'
+        print("\n---------------------------")
+        print("TESTING DATALOADING MODULE")
+        print("the process will generetate two additional folders")
+        print("---------------------------")
+        if not (os.path.isdir("tmpsave")):
+            os.mkdir("tmpsave")
+        cls.eegpath = "Simulated_EEG"
         cls.create_dataset(cls)
-        cls.freq     = 128        # sampling frequency in [Hz]
-        cls.overlap  = 0.15       # overlap between partitions
-        cls.window   = 2          # window length in [seconds]
-
+        cls.freq = 128  # sampling frequency in [Hz]
+        cls.overlap = 0.15  # overlap between partitions
+        cls.window = 2  # window length in [seconds]
 
     def setUp(self):
-        self.seed=1234
+        self.seed = 1234
         random.seed(self.seed)
[…]
