+++ test/EEGself/losses/losses_test.py
@@ -7,195 +7,263 @@ import torch
 from selfeeg import losses
 
 
-
 class TestLoss(unittest.TestCase):
 
     def makeGrid(self, pars_dict):
-        keys=pars_dict.keys()
-        combinations=itertools.product(*pars_dict.values())
-        ds=[dict(zip(keys,cc)) for cc in combinations]
+        keys = pars_dict.keys()
+        combinations = itertools.product(*pars_dict.values())
+        ds = [dict(zip(keys, cc)) for cc in combinations]
         return ds
 
     @classmethod
     def setUpClass(cls):
-        cls.device= torch.device("mps") if torch.backends.mps.is_available() else torch.device("cpu")
-        if cls.device.type == 'cpu':
-            cls.device  = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
+        cls.device = (
+            torch.device("mps") if torch.backends.mps.is_available() else torch.device("cpu")
+        )
+        if cls.device.type == "cpu":
+            cls.device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
         device = cls.device
-        print('\n---------------------')
-        print('TESTING LOSSES MODULE')
-        if cls.device.type != 'cpu':
-            print('Found other device: testing module with both cpu and gpu')
+        print("\n---------------------")
+        print("TESTING LOSSES MODULE")
+        if cls.device.type != "cpu":
+            print("Found other device: testing module with both cpu and gpu")
         else:
-            print('Didn\'t found cuda device: testing module with only cpu')
-        print('---------------------')
+            print("Didn't found cuda device: testing module with only cpu")
+        print("---------------------")
         N, Feat = 64, 128
-        cls.N       = 64
-        cls.Feat    = 128
-        cls.x  = torch.randn(N,Feat)
-        cls.y  = torch.randn(N,Feat)
-        cls.p  = torch.randn(N,Feat)
-        cls.z  = torch.randn(N,Feat)
-        cls.u  = torch.randn(Feat, 1024)
+        cls.N = 64
+        cls.Feat = 128
+        cls.x = torch.randn(N, Feat)
+        cls.y = torch.randn(N, Feat)
+        cls.p = torch.randn(N, Feat)
+        cls.z = torch.randn(N, Feat)
+        cls.u = torch.randn(Feat, 1024)
 
-        if device.type != 'cpu':
-            cls.x2  = torch.randn(N,Feat).to(device=device)
-            cls.y2  = torch.randn(N,Feat).to(device=device)
-            cls.p2  = torch.randn(N,Feat).to(device=device)
-            cls.z2  = torch.randn(N,Feat).to(device=device)
-            cls.u2  = torch.randn(Feat, 1024).to(device=device)
+        if device.type != "cpu":
+            cls.x2 = torch.randn(N, Feat).to(device=device)
+            cls.y2 = torch.randn(N, Feat).to(device=device)
+            cls.p2 = torch.randn(N, Feat).to(device=device)
+            cls.z2 = torch.randn(N, Feat).to(device=device)
+            cls.u2 = torch.randn(Feat, 1024).to(device=device)
 
     def setUp(self):
-        self.seed=1234
+        self.seed = 1234
         np.random.seed(self.seed)
         torch.manual_seed(self.seed)
 
-
     def test_barlowLoss(self):
-        print('Testing Barlow Loss...', end="", flush=True)
[…]
