+++ selfeeg/losses/__init__.py
@@ -1,2 +1,1 @@
-from .losses import (SimCLR_loss, SimSiam_loss, Moco_loss,
-                     BYOL_loss, Barlow_loss, VICReg_loss)
+from .losses import SimCLR_loss, SimSiam_loss, Moco_loss, BYOL_loss, Barlow_loss, VICReg_loss

[…]
