+++ selfeeg/augmentation/functional.py
@@ -1,22 +1,15 @@
 from __future__ import annotations
+
 import math
 import random
-import inspect
+from typing import List, Optional, Union
 
 import numpy as np
-import pandas as pd
-
-from scipy.io import loadmat
-from scipy import signal
-from scipy import interpolate
-from scipy import fft
-
 import torch
 import torch.nn.functional as F
-from torchaudio.functional import lfilter, filtfilt
-
-from typing import Any, Callable, Iterable, TypeVar, Generic, Sequence, List, Dict, Optional, Union
 from numpy.typing import ArrayLike
+from scipy import fft, interpolate, signal
+from torchaudio.functional import filtfilt
 
 from ..utils.utils import torch_pchip
 
