+++ selfeeg/augmentation/__init__.py
@@ -2,35 +2,34 @@
 This is the import of the data augmentation module
 """
 
+from .compose import DynamicSingleAug, RandomAug, SequentialAug, StaticSingleAug
 from .functional import (
-    identity,
-    shift_vertical,
-    shift_horizontal,
-    shift_frequency,
-    flip_vertical,
-    flip_horizontal,
-    scaling,
-    random_slope_scale,
-    random_FT_phase,
-    add_gaussian_noise,
-    add_noise_SNR,
     add_band_noise,
     add_eeg_artifact,
-    get_filter_coeff,
-    moving_avg,
-    filter_lowpass,
-    filter_highpass,
+    add_gaussian_noise,
+    add_noise_SNR,
+    change_ref,
+    channel_dropout,
+    crop_and_resize,
     filter_bandpass,
     filter_bandstop,
-    get_eeg_channel_network_names,
+    filter_highpass,
+    filter_lowpass,
+    flip_horizontal,
+    flip_vertical,
     get_channel_map_and_networks,
-    permute_channels,
-    permutation_signal,
-    warp_signal,
-    crop_and_resize,
-    change_ref,
+    get_eeg_channel_network_names,
+    get_filter_coeff,
+    identity,
     masking,
-    channel_dropout,
+    moving_avg,
+    permutation_signal,
+    permute_channels,
+    random_FT_phase,
+    random_slope_scale,
+    scaling,
+    shift_frequency,
+    shift_horizontal,
+    shift_vertical,
+    warp_signal,
 )
-
-from .compose import StaticSingleAug, DynamicSingleAug, SequentialAug, RandomAug
