+++ selfeeg/utils/__init__.py
@@ -1,10 +1,10 @@
 from .utils import (
+    RangeScaler,
     check_models,
     count_parameters,
-    get_subarray_closest_sum,
-    subarray_closest_sum,
-    scale_range_soft_clip,
-    RangeScaler,
-    torch_pchip,
     create_dataset,
+    get_subarray_closest_sum,
+    scale_range_soft_clip,
+    subarray_closest_sum,
+    torch_pchip,
 )
