+++ selfeeg/models/__init__.py
@@ -1,25 +1,25 @@
 from .zoo import (
+    BasicBlock1,
     ConstrainedConv2d,
     ConstrainedDense,
-    DepthwiseConv2d,
-    SeparableConv2d,
-    EEGNetEncoder,
-    EEGNet,
-    DeepConvNetEncoder,
     DeepConvNet,
-    EEGInceptionEncoder,
+    DeepConvNetEncoder,
+    DepthwiseConv2d,
     EEGInception,
-    TinySleepNetEncoder,
-    TinySleepNet,
-    StagerNetEncoder,
-    StagerNet,
-    ShallowNetEncoder,
-    ShallowNet,
-    BasicBlock1,
-    ResNet1DEncoder,
-    ResNet1D,
-    STNetEncoder,
-    STNet,
-    EEGSymEncoder,
+    EEGInceptionEncoder,
+    EEGNet,
+    EEGNetEncoder,
     EEGSym,
+    EEGSymEncoder,
+    ResNet1D,
+    ResNet1DEncoder,
+    SeparableConv2d,
+    ShallowNet,
+    ShallowNetEncoder,
+    StagerNet,
+    StagerNetEncoder,
+    STNet,
+    STNetEncoder,
+    TinySleepNet,
+    TinySleepNetEncoder,
 )
