### 2024-02-06 MedMaxLab/selfEEG#6 @fedepup
Hi, @vferat. First of all, thank you again for having spent some of your precious time in reviewing selfEEG for JOSS. I believe that your suggestions were really helpful in improving the quality of the library and its understanding.

Here is my final comment related to the extra part about code maintenance. All the things were tested in a forked repository and then merged into this one, so to not mess up things. Please refer to  #7 and #8 for the complete list of changes (number 8 also includes all the pre-commit autofix).

> to add additional tests by installing the "main" version of https://github.com/pytorch/pytorch so as to be able to anticipate the changes needed to remain compatible.

Now tests are run with pytorch nightly instead of with the last stable version

> use a test matrix to check compatibility with several OS/python versions

Now tests are performed on all three os systems and on python 3.10 and 3.11

> to maintain a simple & clean code base, add codestyle tests or a precommit (example: https://github.com/vferat/pycrostates/blob/main/.pre-commit-config.yaml)

I added a pre-commit with basic things I believe are useful for a python project. I'm not used to pre-commit, so for now I decided to add only things of which I completely understood the functionality. As you can see from #8, the code base improved a lot and I was able to find another typo in the dataloading module and lots of unused import (probably related to my initial tests in a debugging notebook). I'm pretty sure I will add other hooks in the future.

I think this will solve everything and this issue can be closed. But I will leave this decision to you.
