### 2024-02-07 #6224 @fedepup
> Thanks to the authors for their responsiveness! The library delivers on its promise to simplify access to EEG data
All my comments have been addressed and I can only recommend to accept this publication.

Thank you very much, @vferat!
