### 2024-02-07 #6224 @fedepup
Hi @emdupre. To help you have a clear view of all the revisions made during **vferat** and **wmvanvliet** review, I would like to write a little summary.

**wmvanvliet:**

1.  **General Checks**:  The copyright holder's name was changed to MedMax Team MedMaxLab/selfEEG#1
2.  **Installation**: Corrected the wrongly reported conda channel MedMaxLab/selfEEG#2
3. **Functionality**: Minor revision to improve compatibility with python < 3.11 MedMaxLab/selfEEG#4
4. **documentation**: Corrected spelling/grammar typos and improved notebooks readability MedMaxLab/selfEEG#3
5. **Paper**:  Corrected spelling/grammar typos MedMaxLab/selfEEG#5 (I also specified name and surname of every author since mine was detected wrongly https://github.com/MedMaxLab/selfEEG/commit/4154e642b52c1d2ae6374c0f71121b2ffe25ebba)

**vferat**

1. **Functionality**: there was a small bug in the dataloading module when assigning float values to the window length or the sampling rate. Corrected and explained here https://github.com/MedMaxLab/selfEEG/issues/6#issuecomment-1904160507
2. **Documentation**: Added a new tutorial notebook with a more concrete use case on real-world data (EEGMMI dataset, preprocessed version) and updated notebooks in the `Notebooks` folder MedMaxLab/selfEEG@e573df17acc1f15a77b0fc94b92aeaa8d23ae3b2
3. **Maintenance (optional but important)**: Improved tests by adding a matrix with different os/python versions and the nighlty version of pytorch MedMaxLab/selfEEG#7  and added a basic list of pre-commit MedMaxLab/selfEEG#8
