### 2024-02-07 #6224 @emdupre
Thank you for your write-up, @fedepup ! I appreciate this summary 🙏 And thank you very much to @wmvanvliet and @vferat for your reviews !

@Bsingstad, please let us know when you've finalized your review -- or if there are any blockers that we should be aware of ! And thank you again for your work on reviewing _SelfEEG_ to date 🌻
