### 2024-02-16 #6224 @fedepup
@emdupre - Thanks for your constant monitoring on this review process. Considering how smoothly things have gone so far and, more importantly, how the revisions have already greatly improved the quality of selfEEG, I'm confident that the review will be completed within the expected time.

@Bsingstad - As already said in the pre-review process, I will try to solve any raised issue as soon as possible, possibly within the same day. Thanks again for having accepted to review selfEEG.
