+++ .github/workflows/doc_build_test.yml
@@ -1,0 +1,35 @@
+name: Sphinx_build_test
+
+on:
+  push:
+    paths:
+      - 'selfeeg/**'
+      - 'docs/**'
+      - '.github/workflows/doc_build_test.yml'
+  pull_request:
+    branches:
+      - main
+    paths:
+      - 'selfeeg/**'
+      - 'docs/**'
+      - '.github/workflows/doc_build_test.yml'
+
+jobs:
+  sphinxbuild:
+
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+          cache: 'pip'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -U sphinx
+          pip install -r docs/doc_requirements.txt
+      - name: Sphinx build
+        run: sphinx-build -M html docs docs
