+++ .github/workflows/doc_build_test.yml
@@ -30,6 +30,8 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -U sphinx
+          pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
+          pip install --upgrade pandoc
           pip install -r docs/doc_requirements.txt
       - name: Sphinx build
         run: sphinx-build -M html docs docs
