+++ .github/workflows/doc_build_test.yml
@@ -33,5 +33,7 @@ jobs:
           pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
           pip install pypandoc_binary
           pip install -r docs/doc_requirements.txt
+      - name: Pandoc
+        uses: docker://pandoc/core:2.9
       - name: Sphinx build
         run: sphinx-build -M html docs docs
