+++ .github/workflows/doc_build_test.yml
@@ -33,11 +33,9 @@ jobs:
           pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
           pip install pypandoc_binary
           pip install pypandoc
-          pip install pandoc
+          # pip install pandoc
           pip install -r docs/doc_requirements.txt
-      - name: Pandoc
-        uses: docker://pandoc/core:2.9
-        with:
-          args: "--help"
+      - name: Pandoc2
+        run: sudo apt-get install pandoc
       - name: Sphinx build
         run: sphinx-build -M html docs docs
