+++ .github/workflows/doc_build_test.yml
@@ -31,11 +31,8 @@ jobs:
           python -m pip install --upgrade pip
           pip install -U sphinx
           pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
-          pip install pypandoc_binary
-          pip install pypandoc
-          # pip install pandoc
           pip install -r docs/doc_requirements.txt
-      - name: Pandoc2
+      - name: Install Pandoc
         run: sudo apt-get install pandoc
       - name: Sphinx build
         run: sphinx-build -M html docs docs
