+++ .github/workflows/python-app.yml
@@ -14,6 +14,10 @@ on:
       - 'test/**'
       - '.github/workflows/python-app.yml'
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
+
 jobs:
   selfeegtest:
 
