### 2024-02-27 #6224 @emdupre
👋 Hi everyone !

I just wanted to follow-up on this review for_SelfEEG_. Given that we're now outside of the six week review window, we'll need to wrap up our initial reviews. @Bsingstad, if you're able to provide comments by **Monday, 4 March** we would still be consider those. Otherwise, we'll need to conclude the initial review.

Thank you again to everyone for all of your input throughout this review process !
