### 2024-03-06 #6224 @emdupre
Hi again,

Now that we've passed this final deadline, I'll need to remove @Bsingstad from the reviewer list. Thank you, though, for your enthusiasm for _SelfEEG_.

As @wmvanvliet and @vferat have already confirmed their completed reviews, I'll then proceed to perform a few final editorial checks.

If there are any questions or concerns at this point, please let me know !
