### 2024-03-12 c7475ba
added torchEEG reference to JOSS paper
