+++ Paper/bibliography.bib
@@ -79,3 +79,14 @@ url = {https://www.mdpi.com/2076-3417/13/9/5472},
 issn = {2076-3417},
 doi = {10.3390/app13095472}
 }
+
+
+@article{torchEEG,
+    title = {{TorchEEGEMO}: A deep learning toolbox towards {EEG}-based emotion recognition},
+	author = {Zhi Zhang and Sheng-hua Zhong and Yan Liu},
+    journal = {Expert Systems with Applications},
+    pages = {123550},
+    year = {2024},
+    issn = {0957-4174},
+	doi = {10.1016/j.eswa.2024.123550}
+}
