### 2024-03-12 #6224 @fedepup
Hi @emdupre.
Yesterday a [paper on torchEEG](https://doi.org/10.1016/j.eswa.2024.123550) was published on _Expert Systems with Applications_. Therefore, I've updated the paper and added the new reference. You can find all the modifications here MedMaxLab/SelfEEG@c7475baa2b539901d9b426ccbde2e0d5038b17c3.

I also take the chance to ask if there are other additional steps I need to perform at this stage of the peer-review.
