### 2024-03-13 ebc8dcd
Update citation formatting in-text
