### 2024-03-13 #6224 @emdupre
Thanks for checking in, @fedepup ! I've now had a chance to review _SelfEEG_ and am happy with the state of the submission 🚀

I do, though, have a few small editorial requests before we can proceed (the _torchEEG_ citation was actually one of my original requests, so thanks for taking care of that !):

- For the author affiliations, we usually recommend that all affiliations follow "Department, Institute/Univ., City, Country" as closely as possible. It looks like additional address information is included; e.g., "Via Gradenigo 6/b, 35131." Could you please remove this additional information or confirm why it is included?
- On line 26, the capitalization of "Artificial Intelligence" is unnecessary
- [ ] Could you please convert the LightlySSL hyperlink on L33 to [their requested citation](https://github.com/lightly-ai/lightly?tab=readme-ov-file#bibtex) as listed in their README? I do not, though, think that their provided article type will render appropriately in the JOSS template. Instead I'd suggest formatting it as follows:
   ```
   @software{susmelj2020lightly,
      title = {Lightly},
      author={Igor Susmelj and Matthias Heller and Philipp Wirth and Jeremy Prescott and Malte Ebner et al.},
      url = {https://github.com/lightly-ai/lightly},
      version = {1.5.0},
      date = {2024-03-11},
   }
   ```
   I've taken the version and date accessed as latest, but please update as appropriate !
- [ ] Similarly, could you please convert the ViSSL hyperlink on the same line to [their requested citation](https://github.com/facebookresearch/vissl?tab=readme-ov-file#citing-vissl) ?
- [ ] Otherwise, there are a few citation formatting issues in the current text. I've [created a pull request](https://github.com/MedMaxLab/selfEEG/pull/12) to update these.
