+++ Paper/bibliography.bib
@@ -46,7 +46,8 @@
 
 @article{pytorch,
   title={Pytorch: An imperative style, high-performance deep learning library},
-  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
+  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, 
+          Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
   journal={Advances in neural information processing systems},
   volume={32},
   year={2019},
@@ -55,38 +56,58 @@
 
 
 @article{eegdl,
-	title     = {A Novel Approach of Decoding EEG Four-class Motor Imagery Tasks via Scout {ESI} and {CNN}},
-	author    = {Hou, Yimin and Zhou, Lu and Jia, Shuyue and Lun, Xiangmin},
-	journal   = {Journal of Neural Engineering},
-	volume    = {17},
-	number    = {1},
-	pages     = {016048},
-	year      = {Feb. 2020},
-	publisher = {IOP Publishing},
-	doi       = {10.1088/1741-2552/ab4af6}
+  title = {A Novel Approach of Decoding EEG Four-class Motor Imagery Tasks via Scout {ESI} and {CNN}},
+  author = {Hou, Yimin and Zhou, Lu and Jia, Shuyue and Lun, Xiangmin},
+  journal = {Journal of Neural Engineering},
+  volume = {17},
+  number = {1},
+  pages = {016048},
+  year = {Feb. 2020},
+  publisher = {IOP Publishing},
+  doi = {10.1088/1741-2552/ab4af6}
 }
 
 
 @Article{app13095472,
-author = {Tshimanga, Louis Fabrice and Del Pup, Federico and Corbetta, Maurizio and Atzori, Manfredo},
-title = {An Overview of Open Source Deep Learning-Based Libraries for Neuroscience},
-journal = {Applied Sciences},
-volume = {13},
-year = {2023},
-number = {9},
-article-number = {5472},
-url = {https://www.mdpi.com/2076-3417/13/9/5472},
-issn = {2076-3417},
-doi = {10.3390/app13095472}
+  author = {Tshimanga, Louis Fabrice and Del Pup, Federico and Corbetta, Maurizio and Atzori, Manfredo},
+  title = {An Overview of Open Source Deep Learning-Based Libraries for Neuroscience},
+  journal = {Applied Sciences},
+  volume = {13},
+  year = {2023},
+  number = {9},
+  article-number = {5472},
+  url = {https://www.mdpi.com/2076-3417/13/9/5472},
+  issn = {2076-3417},
+  doi = {10.3390/app13095472}
 }
 
 
 @article{torchEEG,
-    title = {{TorchEEGEMO}: A deep learning toolbox towards {EEG}-based emotion recognition},
-	author = {Zhi Zhang and Sheng-hua Zhong and Yan Liu},
-    journal = {Expert Systems with Applications},
-    pages = {123550},
-    year = {2024},
-    issn = {0957-4174},
-	doi = {10.1016/j.eswa.2024.123550}
+  title = {{TorchEEGEMO}: A deep learning toolbox towards {EEG}-based emotion recognition},
+  author = {Zhi Zhang and Sheng-hua Zhong and Yan Liu},
+  journal = {Expert Systems with Applications},
+  pages = {123550},
+  year = {2024},
+  issn = {0957-4174},
+  doi = {10.1016/j.eswa.2024.123550}
+}
[…]
