### 2024-03-13 5ae4e47
Corrected typo on JOSS paper
