### 2024-03-13 #6224 @fedepup
HI @emdupre. Here I report all the requested information and revisions.

### Paper

- All the additional address information were removed following the suggested structure ( Department, Institute/Univ., City, Country).
- Capitalization of "Artificial Intelligence" was removed.
- Lightly SSL and ViSSL are now cited with bibtex (I used software instead of misc as you suggested). I also checked that no double parentheses were included in the new draft.

### Tagged release and repository archive

- **_Version Number_**: [v0.1.1](https://github.com/MedMaxLab/selfEEG/releases/tag/v0.1.1)
- _**Zenodo DOI**_:  [10.5281/zenodo.10813095](https://doi.org/10.5281/zenodo.10813095)

Title, list of authors, license (MIT), affiliations and version tag reported in the archive are the same as in the paper or GitHub repository.

### Additional Notes

Zenodo's DOI registration is currently affected by a service incident at DataCite. So, the provided link does not work at the moment, although it is correct. If you check the DOI that represent all versions of selfEEG ( https://zenodo.org/doi/10.5281/zenodo.10813094 ), you can see that this works. You can also check that the one provided for the specific release (v0.1.1) is correct.

Thanks again.
