+++ conda_recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "selfeeg" %}
-{% set version = "0.1.0" %}
+{% set version = "0.1.1" %}
 
 package:
   name: {{ name|lower }}
