### 2024-03-13 df9201a
updated conda recipe with correct SHA256
