### 2024-03-14 #6224 @emdupre
Thank you, @fedepup !

As you can see, I'm having a little difficulty setting the archive with a non-valid DOI. I hope / expect that this will be resolved soon on the DataCite side, but in the meantime I'll try to confirm what we can do on our end.
