### 2024-03-15 #6224 @fedepup
Hi @emdupre. The SelEEG v0.1.1 DOI on Zenodo now works.

Il Gio 14 Mar 2024, 06:02 Elizabeth DuPre ***@***.***> ha
scritto:

> Thank you, @fedepup <https://github.com/fedepup> !
>
> As you can see, I'm having a little difficulty setting the archive with a
> non-valid DOI. I hope / expect that this will be resolved soon on the
> DataCite side, but in the meantime I'll try to confirm what we can do on
> our end.
>
