### 2024-03-15 #6224 @emdupre
Thank you, @fedepup ! I'm now happy to recommend _SelfEEG_ to the EiC team for publication in _JOSS_ ✨🚀 And my personal congratulations on this impressive work !

Thank you, too, to @wmvanvliet and @vferat for your reviews ! _JOSS_ works because of your work 💐
